lc-map_music/requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.2.0
langgraph>=0.2.0
streamlit>=1.37.0
selenium>=4.23.0
webdriver-manager>=4.0.0
python-dotenv>=1.0.1
requests>=2.32.3
pydantic>=2.8.2
pychrome>=0.2.4
